refactor(chain): simplify name-hash resolver layers
